Added KFview, a Shiny app that displays the results from the Kalman processor. This directory should be installed at the root location for the shiny-server (/srv/shiny-server), or it can be run from RStudio while in the KalmanFilter directory.
The KalmanFilter.R script has an added argument for specifying the directory to which to write the output. On datavis, only "KFoutput" works so that will be set there. For other runs, including on EOL outside the Shiny app, the default is to place the output in a subdirectory named "KF" in the normal project directory; this can be changed via the new argument to the script.